test: use vitest instead of the abandoned mocha
