+++ package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "build": "tsc",
     "watch": "tsc -w",
-    "test": "ts-mocha --reporter spec test/*-test.ts"
+    "test": "vitest run"
   },
   "repository": {
     "type": "git",
@@ -27,10 +27,9 @@
   },
   "homepage": "https://github.com/indutny/macho",
   "devDependencies": {
-    "@types/mocha": "*",
     "@types/node": "^20.3.2",
-    "ts-mocha": "*",
-    "typescript": "^5.1.3"
+    "typescript": "^5.1.3",
+    "vitest": "^0.34.3"
   },
   "engines": {
     "node": ">=16",
